requirements.txt
CouchDB==1.2
didkit==0.3.2
ecdsa==0.18.0
Flask==2.2.3
Flask_Cors==3.0.10
Flask_QRcode==3.1.0
Flask_SocketIO==5.3.3
hexbytes==0.3.0
jwcrypto==1.4.2
pytest==7.3.1
python-dotenv==1.0.0
redis==4.5.1
web3==6.0.0
